+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/handler/adapter/ManualHandlerAdapter.java
@@ -1,7 +1,7 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc.handler.adapter;
 
 import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
+import com.interface21.webmvc.servlet.mvc.handler.Controller;
 import com.interface21.webmvc.servlet.view.JspView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -10,7 +10,7 @@ public class ManualHandlerAdapter implements HandlerAdapter{
 
     @Override
     public boolean supports(final Object handler) {
-        return handler instanceof com.interface21.webmvc.servlet.mvc.asis.Controller;
+        return handler instanceof Controller;
     }
 
     @Override
