+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/handler/mapping/HandlerMapping.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc.handler.mapping;
 
 import jakarta.servlet.http.HttpServletRequest;
